Move test rule service + interface to rules package
