+++ src/main/java/com/ericsson/ei/rules/IRuleTestService.java
@@ -1,4 +1,4 @@
-package com.ericsson.ei.services;
+package com.ericsson.ei.rules;
 
 import java.io.IOException;
